change app.js
